Disease-prediction2/requirements.txt

django>=3.2
joblib>=1.1.0
scikit-learn>=1.0.2
psycopg2-binary>=2.9.3
djangorestframework
django-cors-headers
djangorestframework-simplejwt
